requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.121.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.7.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyPDF2==3.0.1
pypdfium2==5.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
